fix: rename customer create-update types file
- It was wrongly named customer.types
