+++ spa/src/root-reducer.ts
@@ -1,10 +1,10 @@
 import { combineReducers } from 'redux';
 import { User } from './modules/auth/user';
 import { Customer } from './modules/crm/customer';
+import { CreateCustomerStepsModel } from './modules/crm/create-update/create-update.types';
 import authReducer from './modules/auth/auth.state';
 import customersReducer from './modules/crm/customers.state';
 import createUpdateCustomersReducer from './modules/crm/create-update/create-update.state';
-import { CreateCustomerStepsModel } from './modules/crm/create-update/customers.types';
 
 export interface StoreState {
   customers: Customer[];
